+++ lib/kettle/rb.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "rb/version"
+require_relative "rb/compat_matrix"
 
 module Kettle
   module Rb